[FE-1192] Moved the ChangeReportModal and ReportsListModal to src/pages/../components
