--- src/components/reportBuilder/index.js
@@ -1,2 +1,0 @@
-export { default as ReportsListModal } from './ReportsListModal';
-export { default as ChangeReportModal } from './ChangeReportModal';
